refactor: change to workspaces structure
This enables separation between server and client dependencies, which
makes it easy to know what is used where.

It also reduces the footprint of the final Docker image, as we only need
the dev deps when building the client bundle, so in the final image we
can include only the server runtime deps.